% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sibship_visibility.R
\name{add_esc_ind_vis}
\alias{add_esc_ind_vis}
\title{add individual visibility based on sib reports to ego X sib X cell reports}
\usage{
add_esc_ind_vis(
  esc.dat,
  ego.id,
  sib.dat,
  sib.frame.indicator,
  varname = "ind_vis"
)
}
\arguments{
\item{esc.dat}{Dataset with a row for each respondent X sibling X cell, likely produced by \code{\link{get_esc_reports}}}

\item{ego.id}{String with the name of the column in \code{esc.dat} containing the survey respondent's id}

\item{sib.dat}{Dataset with a row for each reported sibling, likely produced by TODO}

\item{sib.frame.indicator}{String with the name of the column in \code{sib.dat} containing a 0/1 coded variable indicating whether or not each sib is in the frame population}

\item{varname}{String with the name of the new column to be added to \code{sib.dat} containing the individual visibility}
}
\value{
the ESC dataframe with additional columns that have information about the sibship size and visibility:
\itemize{
\item \code{ind_vis_weight} - the individual visibility weight
\item \code{y.F} - number of reported sibs in frame popn (not including respondent)
\item \code{sibship.size} - total size of reported sibship (including respondent)
}
}
\description{
Takes a dataframe that has a row for each respondent X sib X cell
and adds individual visibility to it
}
\examples{
  # TODO - add example
}
