% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.knownpop.dat}
\alias{example.knownpop.dat}
\title{Example known population data}
\format{
A data frame with 22 rows and 2 variables:
\describe{
\item{known.popn}{The name of the group}
\item{size}{The number of people in the group}
}
}
\usage{
example.knownpop.dat
}
\description{
Example of a household survey dataset, used in unit tests
and vignettes for the \code{networkreporting} package.
}
\keyword{datasets}
