% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.survey}
\alias{example.survey}
\title{Example household survey data}
\format{
A data frame with 2,406 rows and 36 variables:
\describe{
\item{id}{a unique identifier}
\item{cluster}{the cluster (part of the complex survey design)}
\item{region}{the region (part of the complex survey design)}
\item{indweight}{the individual weight (relative)}
\item{sex}{the sex of the respondent}
\item{age.cat}{the age category of the respondent}
\item{widower}{reported connections to widowers}
\item{nurse.or.doctor }{reported connections to nurses or doctors}
\item{male.community.health}{reported connections to male community health workers}
\item{teacher}{reported connections to teachers}
\item{woman.smoke}{reported connections to women who smoke}
\item{priest}{reported connections to priests}
\item{civil.servant }{reported connections to civil servants}
\item{woman.gave.birth}{reported connections to women who gave birth}
\item{muslim}{reported connections to Muslims}
\item{incarcerated}{reported connections to people who are incarcerated}
\item{judge}{reported connections to people who are judges}
\item{man.divorced}{reported connections to men who are divorced}
\item{treatedfortb}{reported connections to men who are divorced}
\item{nsengimana}{reported connections to Nsegimanas}
\item{murekatete}{reported connections to Murekatetes}
\item{twahirwa}{reported connections to Twahirwas}
\item{mukandekezi}{reported connections to Mukandekezis}
\item{nsabimana}{reported connections to Nsabimanas}
\item{mukamana}{reported connections to Mukamanas}
\item{ndayambaje}{reported connections to Ndayambajes}
\item{nyiraneza}{reported connections to Nyiranezas}
\item{bizimana}{reported connections to Bizimanas}
\item{nyirahabimana}{reported connections to Nyirahabimanas}
\item{ndagijimana}{reported connections to Ndagijimanas}
\item{mukandayisenga}{reported connections to Mukandayisengas}
\item{died}{reported connections to people who died}
\item{sex.workers}{reported connections to sex workers}
\item{msm}{reported connections to msm}
\item{idu}{reported connections to injecting drug users}
\item{clients}{reported connections to people who are clients of sex workers}
}
}
\usage{
example.survey
}
\description{
Example of a household survey dataset, used in unit tests
and vignettes for the \code{networkreporting} package.
}
\keyword{datasets}
