% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ec_reports.R
\name{get_ec_reports}
\alias{get_ec_reports}
\title{get ego X cell reports}
\usage{
get_ec_reports(
  esc.dat,
  ego.id,
  sib.dat,
  sib.frame.indicator,
  cell.vars,
  weights,
  ind.vis.var = NULL,
  discretize.exp = FALSE,
  check.frame.consistency = TRUE
)
}
\arguments{
\item{esc.dat}{Dataset with a row for each respondent X sibling X cell, likely produced by \code{\link{get_esc_reports}}}

\item{ego.id}{String with the name of the column in \code{esc.dat} containing the survey respondent's id}

\item{sib.dat}{Dataset with a row for each reported sibling, likely produced by TODO}

\item{sib.frame.indicator}{String with the name of the column in \code{sib.dat} containing a 0/1 coded variable indicating whether or not each sib is in the frame population}

\item{cell.vars}{a vector with Strings containing the names of columns in \code{esc.dat} identifying the cells to group reports by (typically age group, sex, time period)}

\item{weights}{String with the name of the column in \code{esc.dat} that has the survey weights}

\item{ind.vis.var}{String with the name of the column in \code{esc.dat} that has the individual visibility for each sibling; defaults to NULL. If NULL, the individual visibilities are calculated}

\item{discretize.exp}{Boolean for whether or not expsoure should be discretized. Not yet implemented.}

\item{check.frame.consistency}{Boolean; if TRUE (the default), warn when any
occurrence is attributed to an alter marked as being in the frame
population. For mortality that is impossible and indicates a miscoded
frame indicator. Set FALSE when the event being counted is not death.}
}
\value{
A dataframe that has a row for each respondent X cell containing respondent's reported deaths and exposure among siblings in the cell.
}
\description{
Produces a dataframe that has a row for each respondent X cell containing respondent's reported deaths and exposure among siblings in the cell.
}
\examples{
  # TODO - add example
}
