% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goc}
\alias{goc}
\title{Example game of contacts}
\format{a list with some stuff}
\usage{
goc
}
\description{
Example of a game of contacts dataset, collected via
RDS; used in unit tests and vignettes for the \code{networkreporting} 
package.
}
\keyword{datasets}

