\docType{data}
\name{hm.q}
\alias{hm.q}
\title{Names of ARD questions for known populations in example household survey data}
\description{
  A vector containing the names of the columns of which
  correspond to aggregate relational data (ARD) questions
  asked about groups of known size in \code{hhsurvey.data}
}
\keyword{datasets}

