% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_config.R
\name{is_tie_config}
\alias{is_tie_config}
\title{Is this a tie_config?}
\usage{
is_tie_config(x)
}
\arguments{
\item{x}{any object}
}
\value{
\code{TRUE} if \code{x} is a \code{tie_config}
}
\description{
Is this a tie_config?
}
