% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility_rules.R
\name{is_visibility_rule}
\alias{is_visibility_rule}
\title{Is this object a visibility rule?}
\usage{
is_visibility_rule(x)
}
\arguments{
\item{x}{object to test}
}
\value{
\code{TRUE} if \code{x} is a \code{visibility_rule}
}
\description{
Is this object a visibility rule?
}
