% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkreporting_package.r
\docType{data}
\name{MU284}
\alias{MU284}
\title{MU284 population}
\description{
Data used in unit tests for variance estimation.
See TODO-Sarndal TODO-sampling package
TODO-doc describing unit tests
}

