% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkreporting_package.r
\docType{package}
\name{networkreporting-package}
\alias{networkreporting}
\alias{networkreporting-package}
\alias{package-networkreporting}
\title{Network reporting estimators}
\description{
\code{networkreporting} has methods for analyzing data that were collected
using network reporting techniques. It includes estimators appropriate for
indirect sampling, network scale-up, network reporting, and sibling history
methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://dennisfeehan.org/networkreporting/}
}

}
\author{
\strong{Maintainer}: Dennis M. Feehan \email{feehan@berkeley.edu}

Other contributors:
\itemize{
  \item Matthew J. Salganik \email{mjs3@princeton.edu} [thesis advisor]
}

}
