\docType{package}
\name{networkreporting}
\alias{networkreporting}
\alias{networkreporting-package}
\alias{package-networkreporting}
\title{Network reporting estimators}
\description{
  \code{networkreporting} has methods for analyzing data
  that were collected using network reporting techniques.
  It includes estimators appropriate for indirect sampling,
  network scale-up, network reporting, and sibling history
  methods.
}

