\docType{package}
\name{networksampling}
\alias{networksampling}
\alias{networksampling-package}
\alias{package-networksampling}
\title{Network sampling estimators}
\description{
  \code{networksampling} has methods for analyzing data
  that were collected using network sampling techniques. It
  includes estimators appropriate for indirect sampling,
  network scale-up, network reporting, and sibling history
  methods.
}

