\docType{data}
\name{parent.data}
\alias{parent.data}
\title{Example game of contacts parents (nominators)}
\description{
  Example of a game of contacts dataset, collected via RDS,
  organized from the perspective of the parents (ie, those
  who nominated subsequent respondents). See also
  \code{survey.data}. Used in unit tests and vignettes for
  the \code{networkreporting} package.
}
\keyword{datasets}

