\docType{data}
\name{survey.data}
\alias{survey.data}
\title{Example game of contacts survey data}
\description{
  Example of a game of contacts survey dataset, collected
  via RDS. See also \code{parent.data}. Used in unit tests
  and vignettes for the \code{networkreporting} package.
}
\keyword{datasets}

