\docType{data}
\name{target.q}
\alias{target.q}
\title{Names of ARD questions for hidden populations in example household survey data}
\description{
  A vector containing the names of the columns of which
  correspond to aggregate relational data (ARD) questions
  asked about hidden groups (ie, groups of unknown size) in
  \code{hhsurvey.data}
}
\keyword{datasets}

