% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ties.R
\name{tie_cell_vars}
\alias{tie_cell_vars}
\title{Which columns define a cell, given a set of estimate tables?}
\usage{
tie_cell_vars(tabs)
}
\arguments{
\item{tabs}{a list of estimate tables}
}
\value{
character vector of cell column names
}
\description{
Internal. The cell columns are whatever the tables share that is not an
estimated quantity.
}
\keyword{internal}
