% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkreporting_package.r
\docType{data}
\name{toynrnetworks}
\alias{toynrnetworks}
\title{Toy network response networks}
\description{
Toy network examples with attributes attached to each
node; useful for unit tests in the \code{networkreporting}
package.
}

