% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility_rules.R
\name{vis_statistic}
\alias{vis_statistic}
\title{Summarise donor visibilities}
\usage{
vis_statistic(x, w, statistic = c("harmonic", "arithmetic", "median"))
}
\arguments{
\item{x}{numeric vector of donor visibility counts}

\item{w}{numeric vector of donor weights}

\item{statistic}{one of \code{"harmonic"}, \code{"arithmetic"}, \code{"median"}}
}
\value{
a single number
}
\description{
Summarise donor visibilities
}
\keyword{internal}
