% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.survey}
\alias{example.survey}
\title{Example household survey data}
\format{An object of class \code{data.frame} with 2406 rows and 36 columns.}
\usage{
example.survey
}
\description{
Example of a household survey dataset, used in unit tests
and vignettes for the \code{networkreporting} package.
}
\keyword{datasets}

