% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkreporting_package.r
\docType{data}
\name{gocdata}
\alias{gocdata}
\title{Example game of contacts}
\description{
Example of a game of contacts dataset, collected via
RDS; used in unit tests
and vignettes for the \code{networkreporting} package.
}

