% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkreporting_package.r
\docType{data}
\name{hhsurvey}
\alias{hhsurvey}
\title{Example household survey data}
\description{
Example of a household survey dataset, used in unit tests
and vignettes for the \code{networkreporting} package.
}

