% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_population.r
\name{kp.estimator_}
\alias{kp.estimator_}
\alias{kp.estimator}
\title{Average personal network size estimates using known population method}
\usage{
kp.estimator_(
  resp.data,
  known.populations,
  weights,
  boot.weights = NULL,
  ego.id = NULL,
  return.boot = FALSE,
  attribute.names = NULL,
  total.kp.size = NULL,
  alter.popn.size = NULL,
  dropmiss = FALSE,
  verbose = TRUE,
  boot.chunk.size = 500
)

kp.estimator(
  resp.data,
  known.populations,
  weights,
  attribute.names = NULL,
  total.kp.size = 1,
  alter.popn.size = NULL,
  dropmiss = FALSE
)
}
\arguments{
\item{resp.data}{the dataframe that has the survey responses}

\item{known.populations}{the names of the columns in \code{resp.data}
that have respondents' reports about connections to known populations}

\item{weights}{the name of a column that has sampling weights}

\item{boot.weights}{Optional dataframe with bootstrap resampled weights. See Details for more info.}

\item{ego.id}{If boot.weights are included, then this is the name of the
column(s) we need to join the bootstrap weights onto the dataset. This is
most often the id of the ego making the reports.}

\item{return.boot}{if TRUE and boot.weights are included, then return the full bootstrapped estiamates and not just the summaries;
this option causes this function to return a list instead of a tibble}

\item{attribute.names}{the names of the columns in \code{resp.data} that
determine the subgroups for which average degree is estimated;
if NULL, then the average over all respondents is estimated}

\item{total.kp.size}{the size of the probe alters; i.e., the sum of the known population
sizes. if NULL, then this is set to 1}

\item{alter.popn.size}{the size of the population of alters; this is most
often the frame population, which is the default if nothing else is
specified; the size of the frame population is taken to be the sum
of the weights over all of resp.data}

\item{dropmiss}{if FALSE, then, for each row, use only the reports about connections to known populations
that have no missingness. This effectively assumes that missing reports are 0.
if TRUE, then only use rows that have no missingness in reported connections to known
populations in estimating degree.
in this case, the sampling weights are rescaled so that the implied total
size of the frame population is not changed.
(see the 'dropmiss' argument to the function report.aggregator_)
future versions may have other options}

\item{verbose}{if TRUE, print information to screen}

\item{boot.chunk.size}{number of bootstrap weight columns to process at a time;
see the \code{boot.chunk.size} argument of \code{\link{report.aggregator_}}.
Affects speed and peak memory only, not the results. Defaults to 500}
}
\value{
the estimated average degree (\code{dbar.Fcell.F}) for respondents in each
of the categories given by \code{attribute.names}
}
\description{
If given \code{attribute.names}, then this function produces
estimated average network sizes given by the groups that are defined by
all combinations of the attributes; otherwise, it estimates the
average personal network size for the entire frame population.
}
\section{Technical note}{

The estimated average degree is
\eqn{(\sum y_{F_\alpha, A} / N_A) \times
     N_F / N_{F_\alpha}}
here, we estimate \eqn{N_F / N_{F_\alpha}} by dividing the
total of all respondents' weights by the sum of
the weights for respondents in each cell \eqn{\alpha}.
}

\section{TODO}{

\itemize{
\item{ make unit tests }
\item{ think about how to elegantly add options for dbar_(P,Q) vs dbar_(Q,P)}
}
}

\section{Details}{

If you want estimated sampling variances, you can pass in a data frame \code{boot.weights}.
\code{boot.weights} is assumed to have a column that is named whatever the \code{ego.id} is,
and then a series of columns named \code{boot_weight_1}, ..., \code{boot_weight_M}.

The two options for missing values are 'ignore' or 'complete.obs'.
'ignore' adds up each respondent's nonmissing reported connections to the known populations,
effectively treating missing reports as 0s.
'complete.obs' only uses responses from respondents who have non-missing values for all of
the known population reports.
}

