% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_survival.r
\name{network.survival.estimator_}
\alias{network.survival.estimator_}
\alias{network.survival.estimator}
\title{network survival estimator}
\usage{
network.survival.estimator_(resp.data, attribute.data, attribute.names,
  known.populations, total.kp.size = 1, weights, attribute.weights,
  dropmiss = NULL, verbose = TRUE)

network.survival.estimator(resp.data, attribute.data, attribute.names,
  known.populations, total.kp.size = 1, weights, attribute.weights,
  verbose = TRUE)
}
\arguments{
\item{resp.data}{the dataframe that has a row for each respondent, with reported
connections to the groups of known size, as well as the attributes.
Note that the column names of the attributes should match
their names in \code{attribute.data}}

\item{attribute.data}{A dataframe with the reported attributes of hidden population 
members reported by survey respondents. There should be one 
row for each time a respondent reports a hidden population member.
For example, to estimate death rates, there should be one row for
each report of a death.}

\item{attribute.names}{the names of the columns of attribute.data
and resp.data that contain the attribute information.}

\item{known.populations}{the names of the columns in \code{resp.data} that
have responses to the known population questions}

\item{total.kp.size}{the size of the probe alters, i.e., the sum of the known
population sizes}

\item{weights}{the weights or weights column for the respondent data}

\item{attribute.weights}{the weights or weights column for the alter data}

\item{dropmiss}{see \code{\link{report.aggregator}}}

\item{verbose}{if TRUE, print information to screen}
}
\value{
the network reporting estimate of the hidden population's size
        (as a prevalence) broken down by the categories defined by all combinations
        of \code{attribute.names}.
}
\description{
use an aggregate multiplicity estimator
and the respondents' own network size estimates
to estimate hidden population sizes
}
\details{
This function takes two sources of data as input: first, it requires
a long-form dataframe with the attributes of the reported members of
the hidden population. For example, if we are asking about emigres and
we collect the age and sex of each reported emigrant, then the long form
dataset might look like:
\tabular{ccc}{
  age \tab sex \tab weight\cr
   15 \tab m   \tab 2.10  \cr
   58 \tab f   \tab 1.15  \cr
   33 \tab m   \tab 3.67  \cr
}
 
 The second source of data we need is the known population responses for the
 respondents, along with the *same* attributes for each respondent. For
 example, in the situation above, we would also require a dataset like this
 to be passed in
 \tabular{cccccc}{
  age  \tab sex  \tab  weight \tab hm.teachers \tab  hm.nurses  \tab ...  \cr
  20   \tab   f  \tab  2.10   \tab 4           \tab  0          \tab ...  \cr
  44   \tab   m  \tab  1.65   \tab 0           \tab  2          \tab ...  \cr
  60   \tab   m  \tab  2.75   \tab 1           \tab  1          \tab ...  \cr
}
}
\section{Technical note}{

This function assumes that the sampling weights are standard analysis weights
and *not* relative weights. Standard analysis weights should provide an estimate
for the size of the frame population when added up; relative weights, on the other hand,
will sum to the number of respondents in the sample.  Demographic and Health surveys
typically have relative weights, which must be converted into standard sampling weights
before using this function.
}

\section{Returned dataframe}{

Currently, the dataframe that is returned has columns for the attributes that death rates
were calculated for, along with:
\itemize{
\item{ \code{sum.y.kp} - from \code{kp.estimator}}
\item{ \code{wgt.total.y.kp} - from \code{kp.estimator}}
\item{ \code{num.obs.y.kp} - from \code{kp.estimator}}
\item{ \code{sum.y.kp.over.kptot} - from \code{kp.estimator}}
\item{ \code{dbar.Fcell.F} - estimated avg. number connections from F.alpha to F}
\item{ \code{mean.deaths} - weighted avg number of reported deaths}
\item{ \code{sum.deaths} - weighted total number of reported deaths}
\item{ \code{wgt.total.deaths} - total of weights (estimate for N.F)}
\item{ \code{wgt.inv.total.deaths} - (not used)}
\item{ \code{num.obs.deaths} - number of observations}
\item{ \code{asdr.hat} - the estimated death rate}
}
}

\section{TODO}{

\itemize{
\item{ handle missing values}
\item{ think about whether or not this is the best way to handle N.F}
\item{ write more general agg mult est fn and call that }
\item{ make unit tests }
}
}

