% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_up.r
\name{plot_meanties_truth}
\alias{plot_meanties_truth}
\title{plot_meanties_truth}
\usage{
plot_meanties_truth(survey.data, weights = NULL, known.popns = NULL)
}
\arguments{
\item{survey.data}{the dataframe with the survey results}

\item{weights}{if not NULL, weights to use in computing the estimate. this
 should be the name of the column in the survey.data which
has the variable with the appropriate weights. these weights
 should be construted so that, eg, the mean of the degrees
 is estimated as (1/n) * \\sum_i {w_i * d_i}}

\item{known.popns}{if not NULL, a vector whose entries are the
size of the known
populations, and whose names are the variable names in the dataset
corresponding to each one. if NULL, then assume that the survey.data
dataframe has an attribute called 'known.popns' containing this vector.}
}
\value{
a ggplot2 object with the relationship plot
}
\description{
plot the relationship between the mean number of
ties in the survey dataset and the true popn sizes
}
\details{
TODO - more in-depth description of this function
}

