% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_aggregator.r
\name{report.aggregator_}
\alias{report.aggregator}
\alias{report.aggregator_}
\title{aggregate a reported quantity by groups}
\usage{
report.aggregator_(resp.data, attribute.names, qoi, weights, qoi.name,
  dropmiss = FALSE)

report.aggregator(resp.data, attribute.names = NULL, qoi, weights,
  qoi.name = NULL, dropmiss = FALSE)
}
\arguments{
\item{resp.data}{the data}

\item{attribute.names}{the names of the variables that define
the groups for which the qoi should be aggregated}

\item{qoi}{the variable with quantity to aggregate}

\item{weights}{analysis weights}

\item{qoi.name}{the name of the qoi}

\item{dropmiss}{NOT YET IMPLEMENTED}
}
\value{
the estimated average degree for respondents in each
        of the categories given by \code{attribute.names}
}
\description{
This function takes a quantity and aggregates it by groups,
using the design weights.
}

